+++ src/main/python/.gitignore
@@ -1,0 +1,1 @@
+__pycache__
